Refactor for cleaner code
