+++ crates/celo-revm/src/api/mod.rs
@@ -1,4 +1,5 @@
 pub mod builder;
 pub mod celo_block_env;
+pub mod celo_system_tx;
 pub mod default_ctx;
 pub mod exec;
